Use double type not decimal for db-migrate

+++ migrations/20210403071712-add-segments.js
@@ -52,7 +52,7 @@ exports.up = function (db) {
     len_m: {
       defaultValue: 0.0,
       notNull: true,
-      type: 'decimal',
+      type: 'double',
     },
     base: {
       notNull: true,
